+++ src/communication/registervaluehandler.h
@@ -18,17 +18,17 @@ public:
     void setRegisters(QList<ModbusRegister> &registerList);
 
     void startRead();
-    void processPartialResult(QMap<quint16, Result> partialResultMap, quint8 connectionId);
+    void processPartialResult(QMap<quint32, Result> partialResultMap, quint8 connectionId);
     void finishRead();
 
-    void registerAddresList(QList<quint16>& registerList, quint8 connectionId);
+    void registerAddresList(QList<quint32>& registerList, quint8 connectionId);
 
 signals:
     void registerDataReady(QList<Result> registers);
 
 private:
 
-    uint32_t convertEndianness(bool bLittleEndian, uint16_t value, uint16_t nextValue);
+    uint32_t convertEndianness(bool bLittleEndian, quint16 value, quint16 nextValue);
 
     SettingsModel* _pSettingsModel;
